move BoltztrapPlotter to plotting package (where dependencies like matplotlib are already there and not mixed into running BTrap)
